pyproject.toml
[tool.poetry]
name = "p"
version = "0.1.0"
description = ""
authors = ["mohser sh <sh1361m@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.11.5"
mlflow = "^2.6.0"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
prefect = "^2.12.0"
prefect-ray = "^0.2.5"
grpcio = "^1.57.0"
ray = "^2.6.3"
matplotlib = "^3.7.2"
wordcloud = "^1.9.2"
nltk = "^3.8.1"
pytest = "^7.4.2"
coverage = "^7.3.1"
pylint = "^2.17.5"
dvc = {extras = ["s3"], version = "^3.19.0"}
dvc-s3 = "^2.23.0"
seaborn = "^0.12.2"
dvclive = "^2.16.0"
evidently = "^0.4.4"



[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
